BUGFIX: Prevent EnumType() accessor vs EnumType.NULL_VALUE name clash when sinceVersion used.
C# uses an upper case notation for both method names and types. When the same name is used for the accessor and the type, it causes the following  name clash.

```
public SomeEnum SomeEnum()
    {
        if (_actingVersion < 1) return SomeEnum.NULL_VALUE;

        return (SomeEnum)unchecked((sbyte)_buffer.GetByte(_offset + 29));
    }
```

This is a compilation error.

The solution is to use a fully qualified name in this case.

```
public SomeEnum SomeEnum()
    {
        if (_actingVersion < 1) return Com.Example.SomeEnum.NULL_VALUE;

        return (SomeEnum)unchecked((sbyte)_buffer.GetByte(_offset + 29));
    }
```

This fixed the problem.